refactor: extract RuntimeStateDoc into runtime-doc crate
Move RuntimeStateDoc, RuntimeStateError, snapshot types, and
StreamOutputState from notebook-doc to the new runtime-doc crate.

JSON/Automerge helpers (read_json_value, insert_json_at_index, etc.)
are duplicated in runtime-doc to avoid a circular dependency. notebook-doc
keeps them pub for its own NotebookDoc use.

notebook-doc/src/runtime_state.rs becomes a re-export shim (pub use
runtime_doc::*) so downstream crates compile unchanged.

266 tests move to runtime-doc. No behavioral changes.
