+++ crates/runtime-doc/Cargo.toml
@@ -10,6 +10,5 @@ license.workspace = true
 automerge = "0.8"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-notebook-doc = { path = "../notebook-doc" }
 tokio = { version = "1", features = ["sync"] }
 tracing = "0.1"
