+++ crates/runtime-doc/src/lib.rs
@@ -1,7 +1,9 @@
-pub mod error;
-
 mod doc;
+pub mod error;
 mod handle;
+pub mod json_helpers;
 mod types;
 
+pub use doc::*;
 pub use error::RuntimeStateError;
+pub use types::*;
