+++ crates/runtime-doc/src/types.rs
@@ -1,1 +1,5 @@
-// Populated in Task 3
+#[derive(Debug, Clone)]
+pub struct StreamOutputState {
+    pub index: usize,
+    pub blob_hash: String,
+}
